requirements.txt
numpy==1.19.2
pandas==1.1.3
pkg-resources==0.0.0
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.20
